requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
distlib==0.3.9
filelock==3.16.1
firebase-admin==6.5.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h5py==3.12.1
httplib2==0.22.0
humanfriendly==10.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
ncnn==1.0.20241226
networkx==3.3
numpy==1.26.4
onnx==1.17.0
onnx-graphsurgeon==0.5.2
onnx2tf==1.26.2
onnxruntime==1.19.2
onnxslim==0.1.34
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.1.1
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sng4onnx==1.0.4
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
tf_keras==2.17.0
tflite-support==0.1.0a1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.5
ultralytics-thop==2.0.9
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.27.0
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.16.0
